userstory-format-agent/requirements.txt
--extra-index-url https://test.pypi.org/simple/

# Core A2A and Server dependencies
a2a-sdk==0.2.11
uvicorn==0.35.0
httpx==0.28.1
fastapi==0.115.13

# Semantic Kernel and Copilot Studio dependencies
semantic-kernel==1.34.0
microsoft-agents-core
microsoft-agents-copilotstudio-client